fix: reorganize some class and update wire. :bug:
