+++ go.mod
@@ -19,9 +19,9 @@ require (
 	github.com/zitadel/oidc/v3 v3.29.1
 	go.mongodb.org/mongo-driver v1.16.1
 	go.uber.org/mock v0.4.0
+	golang.org/x/oauth2 v0.23.0
 	google.golang.org/grpc v1.64.0
 	google.golang.org/protobuf v1.33.0
-	golang.org/x/oauth2 v0.23.0
 )
 
 require (
@@ -43,12 +43,12 @@ require (
 	github.com/swaggo/swag v1.16.3 // indirect
 	github.com/zitadel/logging v0.6.0 // indirect
 	github.com/zitadel/schema v1.3.0 // indirect
-	golang.org/x/net v0.25.0 // indirect
-	google.golang.org/genproto/googleapis/rpc v0.0.0-20240318140521-94a12d6c2237 // indirect
 	go.opentelemetry.io/otel v1.29.0 // indirect
 	go.opentelemetry.io/otel/metric v1.29.0 // indirect
 	go.opentelemetry.io/otel/trace v1.29.0 // indirect
+	golang.org/x/net v0.26.0 // indirect
 	golang.org/x/tools v0.21.1-0.20240508182429-e35e4ccd0d2d // indirect
+	google.golang.org/genproto/googleapis/rpc v0.0.0-20240318140521-94a12d6c2237 // indirect
 )
 
 require (
